Add `Predicate` class (#41)
